+++ docs/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### 10.2.0 Fixes
 
+- `[General]` Temporary Placeholder
+- `[Multiselect]` Added an example showing automation id attributes on options in multiselect. ([#1005](https://github.com/infor-design/enterprise-ng/issues/1005))
+
 ### 10.2.0 Features
 
 - `[Cards]` Added Angular Wrapper Component for the IDS Cards. ([#1036](https://github.com/infor-design/enterprise-ng/issues/1036)) `EA`